Fixed the accesibility for mobile navbarcomponent:
for hamburger menu and the close button I added aria-label attributes, which provide a textual description of the button's purpose to assistive technology users.
for l-inks I added tabIndex attributes that control whether the links are included in the tab order when the menu is open or closed. When the menu is closed, the links are excluded from the tab order (tabIndex={-1}), preventing keyboard users from accidentally navigating to them.